refactor: move frontend visual test machinery to util/frontend_visual_test
Extract the Puppeteer-based visual regression test runner (visual-test-lib.mjs),
hermetic fonts (Inter.woff2, fonts.conf, test-fonts.css), and their Bazel
targets (visual_test_lib, fonts, fonts_conf) from props/frontend/tests/ into a
new shared util/frontend_visual_test/ workspace member.

Update props/frontend/defs.bzl and all visual_test_*.mjs test entry points to
reference //util/frontend_visual_test instead of //props/frontend. Register the
new workspace member in pnpm-workspace.yaml and MODULE.bazel.
