+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - "props/frontend"
   - "agent_server/web"
   - "hack/rspcache/admin_ui"
+  - "util/frontend_visual_test"
