+++ props/frontend/defs.bzl
@@ -4,18 +4,19 @@ load("@aspect_rules_js//js:defs.bzl", "js_test")
 
 # Assets shared by every visual scenario test.
 _VISUAL_BASE_DATA = [
-    "//props/frontend:visual_test_lib",
+    "//util/frontend_visual_test:visual_test_lib",
     "//props/frontend:harness_bundle",
     "//props/frontend:harness_js",
     "//props/frontend:visual_test_assets",
-    "//props/frontend:fonts_conf",
+    "//util/frontend_visual_test:fonts",
+    "//util/frontend_visual_test:fonts_conf",
     "@playwright_browsers//:chromium-headless-shell",
 ]
 
 _VISUAL_ENV = {
     "HARNESS_PATH": "$(rootpath //props/frontend:harness_js)",
     "PUPPETEER_EXECUTABLE_PATH": "$(rootpath @playwright_browsers//:chromium-headless-shell)",
-    "FONTCONFIG_FILE": "$(rootpath //props/frontend:fonts_conf)",
+    "FONTCONFIG_FILE": "$(rootpath //util/frontend_visual_test:fonts_conf)",
     "FREETYPE_PROPERTIES": "cff:no-stem-darkening=1",
 }
