+++ props/frontend/src/components/stats/visual_test_OccurrenceStatsTable.mjs
@@ -1,3 +1,3 @@
-import { main } from "../../../tests/visual-test-lib.mjs";
+import { main } from "../../../../../util/frontend_visual_test/visual-test-lib.mjs";
 
 await main("OccurrenceStatsTable", import.meta.url);
