+++ props/frontend/src/components/visual_test_FileViewerAnnotated.mjs
@@ -1,3 +1,3 @@
-import { main } from "../../tests/visual-test-lib.mjs";
+import { main } from "../../../../util/frontend_visual_test/visual-test-lib.mjs";
 
 await main("FileViewerAnnotated", import.meta.url);
