+++ props/frontend/src/components/visual_test_RunDetailCritic.mjs
@@ -1,3 +1,3 @@
-import { main } from "../../tests/visual-test-lib.mjs";
+import { main } from "../../../../util/frontend_visual_test/visual-test-lib.mjs";
 
 await main("RunDetailCritic", import.meta.url);
