+++ props/frontend/src/pages/visual_test_SnapshotDetail.mjs
@@ -1,3 +1,3 @@
-import { main } from "../../tests/visual-test-lib.mjs";
+import { main } from "../../../../util/frontend_visual_test/visual-test-lib.mjs";
 
 await main("SnapshotDetail", import.meta.url);
