+++ props/frontend/tests/harness/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Component Harness</title>
-    <link rel="stylesheet" href="./test-fonts.css" />
+    <link rel="stylesheet" href="../../../../util/frontend_visual_test/test-fonts.css" />
     <link rel="stylesheet" href="./dist/harness.css" />
   </head>
   <body>
