+++ util/frontend_visual_test/BUILD.bazel
@@ -1,0 +1,48 @@
+load("@aspect_rules_js//js:defs.bzl", "js_library")
+load("@npm_ducktape//:defs.bzl", "npm_link_all_packages")
+
+# Link npm packages for this workspace member
+npm_link_all_packages(name = "node_modules")
+
+# Puppeteer-based visual regression test runner.
+js_library(
+    name = "visual_test_lib",
+    srcs = ["visual-test-lib.mjs"],
+    tags = ["no-lint"],
+    visibility = [
+        "//approval_gate/frontend:__pkg__",
+        "//props/frontend:__pkg__",
+        "//props/frontend/src:__subpackages__",
+    ],
+    deps = [
+        ":node_modules/pixelmatch",
+        ":node_modules/pngjs",
+        ":node_modules/puppeteer",
+    ],
+)
+
+# Font files for hermetic rendering.
+filegroup(
+    name = "fonts",
+    srcs = [
+        "fonts/Inter.woff2",
+        "fonts/fonts.conf",
+        "test-fonts.css",
+    ],
+    visibility = [
+        "//approval_gate/frontend:__pkg__",
+        "//props/frontend:__pkg__",
+        "//props/frontend/src:__subpackages__",
+    ],
+)
+
+# Single-file target for FONTCONFIG_FILE env var (rootpath needs a single file).
+filegroup(
+    name = "fonts_conf",
+    srcs = ["fonts/fonts.conf"],
+    visibility = [
+        "//approval_gate/frontend:__pkg__",
+        "//props/frontend:__pkg__",
+        "//props/frontend/src:__subpackages__",
+    ],
+)
