+++ util/frontend_visual_test/test-fonts.css
@@ -9,7 +9,7 @@
 
 @font-face {
   font-family: "Inter";
-  src: url("../fonts/Inter.woff2") format("woff2");
+  src: url("./fonts/Inter.woff2") format("woff2");
   font-weight: 100 900;
   font-display: block;
 }
